+++ platform/emulator/var_bool.hh
@@ -98,7 +98,7 @@ public:
   // tagging and untagging constrained variables
   //
   OZ_FDIntVar * getTag(void) {
-    return (OZ_FDIntVar *)  (u.var_type & ~u_mask);
+    return (OZ_FDIntVar *)  (uintptr_t) (u.var_type & ~u_mask);
   }
   //
   // end of tagging ...
